thermalcycle: update job template and add submitter script

+++ generators/thermalcycle/make_jscripts.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
 
-# TODO: possibly generalise for different machines
-
 DEBUG=""
 if [ ! -z "$1" ]; then
   DEBUG=1
 fi
+
 
 kappas=(0.139986 0.140026 0.140046 0.140056 0.140066 0.140076 0.140086 0.140106 0.140146 0.140186)
 trajs=(400       400      200      100      100      100      100      200      400      400)
@@ -23,6 +22,7 @@ job_template=jureca.template.job
 basename=thermalcycle
 baseseed=1741726
 
+PRETEND=0
 
 JNAME=iwa_b1.726-L24T48-csw1.74-mul0.0009-musigma0.1408-mudelta0.1521_thermalcycle
 RUNDIR=$WORK/runs/nf211/${JNAME}
@@ -70,27 +70,30 @@ for direction in 0 1; do
 
       if [ ! -z "${DEBUG}" ]; then
         echo kappa=$kappa 2kappamu=$kappa2mu 2kappamubar=$kappa2mubar 2kappamudel=$kappa2mudel traj=${trajs[idx]} njobs=$njobs
+        echo m0= $( echo "1.0 / (2 * $kappa)" | bc -l ) 
         echo jfile=${jfile} 
         echo ifile=${ifile} 
         echo
       fi
+      
+      if [ ! $PRETEND -eq 1 ]; then
+        cp $job_template $jfile
+        sed -i "s@SLJOBNAME@${jidstring}@g" $jfile
+        sed -i "s@SLIFILE@${ifile}@g" $jfile
+        sed -i "s@SLOFILE@${ofile}@g" $jfile
+        sed -i "s@SLWDIR@${WDIR}@g" $jfile
+        if [ ${jobid} -gt 0 -a ${kjobid} -eq 0 ]; then
+          sed -i "s@prevWDIR=@prevWDIR=${prevWDIR}@g" $jfile
+        fi
 
-      cp $job_template $jfile
-      sed -i "s@SLJOBNAME@${jidstring}@g" $jfile
-      sed -i "s@SLIFILE@${ifile}@g" $jfile
-      sed -i "s@SLOFILE@${ofile}@g" $jfile
-      sed -i "s@SLWDIR@${WDIR}@g" $jfile
-      if [ ${jobid} -gt 0 -a ${kjobid} -eq 0 ]; then
-        sed -i "s@prevWDIR=@prevWDIR=${prevWDIR}@g" $jfile
+        cp $input_template $ifile
+        sed -i "s/SEED/${seed}/g" $ifile
+        sed -i "s/NMEAS/${traj_per_job}/g" $ifile
+        sed -i "s/KAPPAONLY/${kappa}/g" $ifile
+        sed -i "s/KAPPAMU2/${kappa2mu}/g" $ifile
+        sed -i "s/KAPPAMUBAR2/${kappa2mubar}/g" $ifile
+        sed -i "s/KAPPAMUDEL2/${kappa2mudel}/g" $ifile
       fi
-
-      cp $input_template $ifile
-      sed -i "s/SEED/${seed}/g" $ifile
-      sed -i "s/NMEAS/${traj_per_job}/g" $ifile
-      sed -i "s/KAPPAONLY/${kappa}/g" $ifile
-      sed -i "s/KAPPAMU2/${kapa2mu}/g" $ifile
-      sed -i "s/KAPPAMUBAR2/${kappa2mubar}/g" $ifile
-      sed -i "s/KAPPAMUDEL2/${kappa2mudel}/g" $ifile
 
       jobid=$(( $jobid + 1 ))
     done
